audio: Clarify "bus" device type
"Bus" devices are permanently attached. That means, the
connection type must be empty, hence `CONNECTION_BUS`
must be deprecated.

Since "bus" devices were initially created as `{IN|OUT}_DEVICE`
type with `CONNECTION_BUS`, which was later changed to an empty
connection, a new alias `{IN|OUT}_BUS` is added for more clear
code.

Bug: 303926380
Test: m
Change-Id: I5ab391e9aace2d3c58f68fdd238621cc61e16afa